Move saveDataCsv to SettingsViewModel
