+++ app/src/main/java/com/geeksville/mesh/ui/settings/SettingsViewModel.kt
@@ -17,37 +17,60 @@
 
 package com.geeksville.mesh.ui.settings
 
+import android.app.Application
+import android.net.Uri
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.geeksville.mesh.IMeshService
 import com.geeksville.mesh.LocalOnlyProtos.LocalConfig
+import com.geeksville.mesh.MeshProtos
+import com.geeksville.mesh.Portnums
+import com.geeksville.mesh.Position
+import com.geeksville.mesh.android.Logging
 import com.geeksville.mesh.android.prefs.UiPrefs
+import com.geeksville.mesh.database.MeshLogRepository
 import com.geeksville.mesh.database.NodeRepository
 import com.geeksville.mesh.database.entity.MyNodeEntity
 import com.geeksville.mesh.model.Node
 import com.geeksville.mesh.repository.datastore.RadioConfigRepository
+import com.geeksville.mesh.util.positionToMeter
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
+import java.io.BufferedWriter
+import java.io.FileNotFoundException
+import java.io.FileWriter
+import java.text.SimpleDateFormat
+import java.util.Locale
 import javax.inject.Inject
+import kotlin.math.roundToInt
 
 @HiltViewModel
 class SettingsViewModel
 @Inject
 constructor(
+    private val app: Application,
     private val radioConfigRepository: RadioConfigRepository,
     private val nodeRepository: NodeRepository,
+    private val meshLogRepository: MeshLogRepository,
     private val uiPrefs: UiPrefs,
-) : ViewModel() {
+) : ViewModel(),
+    Logging {
     val myNodeInfo: StateFlow<MyNodeEntity?>
         get() = nodeRepository.myNodeInfo
+
+    val myNodeNum
+        get() = myNodeInfo.value?.myNodeNum
 
     val ourNodeInfo: StateFlow<Node?>
         get() = nodeRepository.ourNodeInfo
@@ -84,10 +107,144 @@ constructor(
     val excludedModulesUnlocked: StateFlow<Boolean> = _excludedModulesUnlocked.asStateFlow()
 
     fun setProvideLocation(value: Boolean) {
-        myNodeInfo.value?.myNodeNum?.let { uiPrefs.setShouldProvideNodeLocation(it, value) }
+        myNodeNum?.let { uiPrefs.setShouldProvideNodeLocation(it, value) }
     }
 
     fun unlockExcludedModules() {
         viewModelScope.launch { _excludedModulesUnlocked.value = true }
+    }
+
+    /**
+     * Export all persisted packet data to a CSV file at the given URI.
+     *
+     * The CSV will include all packets, or only those matching the given port number if specified. Each row contains:
+     * date, time, sender node number, sender name, sender latitude, sender longitude, receiver latitude, receiver
+     * longitude, receiver elevation, received SNR, distance, hop limit, and payload.
+     *
+     * @param uri The destination URI for the CSV file.
+     * @param filterPortnum If provided, only packets with this port number will be exported.
+     */
+    @Suppress("detekt:CyclomaticComplexMethod", "detekt:LongMethod")
+    fun saveDataCsv(uri: Uri, filterPortnum: Int? = null) {
+        viewModelScope.launch(Dispatchers.Main) {
+            // Extract distances to this device from position messages and put (node,SNR,distance)
+            // in the file_uri
+            val myNodeNum = myNodeNum ?: return@launch
+
+            // Capture the current node value while we're still on main thread
+            val nodes = nodeRepository.nodeDBbyNum.value
+
+            val positionToPos: (MeshProtos.Position?) -> Position? = { meshPosition ->
+                meshPosition?.let { Position(it) }.takeIf { it?.isValid() == true }
+            }
+
+            writeToUri(uri) { writer ->
+                val nodePositions = mutableMapOf<Int, MeshProtos.Position?>()
+
+                @Suppress("MaxLineLength")
+                writer.appendLine(
+                    "\"date\",\"time\",\"from\",\"sender name\",\"sender lat\",\"sender long\",\"rx lat\",\"rx long\",\"rx elevation\",\"rx snr\",\"distance\",\"hop limit\",\"payload\"",
+                )
+
+                // Packets are ordered by time, we keep most recent position of
+                // our device in localNodePosition.
+                val dateFormat = SimpleDateFormat("\"yyyy-MM-dd\",\"HH:mm:ss\"", Locale.getDefault())
+                meshLogRepository.getAllLogsInReceiveOrder(Int.MAX_VALUE).first().forEach { packet ->
+                    // If we get a NodeInfo packet, use it to update our position data (if valid)
+                    packet.nodeInfo?.let { nodeInfo ->
+                        positionToPos.invoke(nodeInfo.position)?.let { nodePositions[nodeInfo.num] = nodeInfo.position }
+                    }
+
+                    packet.meshPacket?.let { proto ->
+                        // If the packet contains position data then use it to update, if valid
+                        packet.position?.let { position ->
+                            positionToPos.invoke(position)?.let {
+                                nodePositions[
+                                    proto.from.takeIf { it != 0 } ?: myNodeNum,
+                                ] = position
+                            }
+                        }
+
+                        // packets must have rxSNR, and optionally match the filter given as a param.
+                        if (
+                            (filterPortnum == null || proto.decoded.portnumValue == filterPortnum) &&
+                            proto.rxSnr != 0.0f
+                        ) {
+                            val rxDateTime = dateFormat.format(packet.received_date)
+                            val rxFrom = proto.from.toUInt()
+                            val senderName = nodes[proto.from]?.user?.longName ?: ""
+
+                            // sender lat & long
+                            val senderPosition = nodePositions[proto.from]
+                            val senderPos = positionToPos.invoke(senderPosition)
+                            val senderLat = senderPos?.latitude ?: ""
+                            val senderLong = senderPos?.longitude ?: ""
+
+                            // rx lat, long, and elevation
+                            val rxPosition = nodePositions[myNodeNum]
+                            val rxPos = positionToPos.invoke(rxPosition)
+                            val rxLat = rxPos?.latitude ?: ""
+                            val rxLong = rxPos?.longitude ?: ""
+                            val rxAlt = rxPos?.altitude ?: ""
+                            val rxSnr = proto.rxSnr
+
+                            // Calculate the distance if both positions are valid
+
+                            val dist =
+                                if (senderPos == null || rxPos == null) {
+                                    ""
+                                } else {
+                                    positionToMeter(
+                                        Position(rxPosition!!), // Use rxPosition but only if rxPos was
+                                        // valid
+                                        Position(senderPosition!!), // Use senderPosition but only if
+                                        // senderPos was valid
+                                    )
+                                        .roundToInt()
+                                        .toString()
+                                }
+
+                            val hopLimit = proto.hopLimit
+
+                            val payload =
+                                when {
+                                    proto.decoded.portnumValue !in
+                                        setOf(
+                                            Portnums.PortNum.TEXT_MESSAGE_APP_VALUE,
+                                            Portnums.PortNum.RANGE_TEST_APP_VALUE,
+                                        ) -> "<${proto.decoded.portnum}>"
+
+                                    proto.hasDecoded() -> proto.decoded.payload.toStringUtf8().replace("\"", "\"\"")
+
+                                    proto.hasEncrypted() -> "${proto.encrypted.size()} encrypted bytes"
+                                    else -> ""
+                                }
+
+                            //  date,time,from,sender name,sender lat,sender long,rx lat,rx long,rx
+                            // elevation,rx
+                            // snr,distance,hop limit,payload
+                            @Suppress("MaxLineLength")
+                            writer.appendLine(
+                                "$rxDateTime,\"$rxFrom\",\"$senderName\",\"$senderLat\",\"$senderLong\",\"$rxLat\",\"$rxLong\",\"$rxAlt\",\"$rxSnr\",\"$dist\",\"$hopLimit\",\"$payload\"",
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    private suspend inline fun writeToUri(uri: Uri, crossinline block: suspend (BufferedWriter) -> Unit) {
+        withContext(Dispatchers.IO) {
+            try {
+                app.contentResolver.openFileDescriptor(uri, "wt")?.use { parcelFileDescriptor ->
+                    FileWriter(parcelFileDescriptor.fileDescriptor).use { fileWriter ->
+                        BufferedWriter(fileWriter).use { writer -> block.invoke(writer) }
+                    }
+                }
+            } catch (ex: FileNotFoundException) {
+                errormsg("Can't write file error: ${ex.message}")
+            }
+        }
     }
 }
